+++ crates/ergot-base/Cargo.toml
@@ -37,10 +37,11 @@ std = [
     "tokio",
 ]
 embassy-usb-v0_4 = [
+    "dep:defmt",
     "dep:embassy-usb",
     "dep:embassy-futures",
     "dep:embassy-time",
-    "dep:defmt",
+    "dep:static_cell",
 ]
 nusb-v0_1 = [
     "std",
@@ -59,13 +60,16 @@ cordyceps       = { version = "0.3.4",  default-features = false }
 maitake-sync    = { version = "0.2.1",  default-features = false }
 mutex           = { version = "1.0.0",  features = ["impl-critical-section"] }
 serde           = { version = "1.0",    features = ["derive"], default-features = false }
+
+# std
 tokio           = { version = "1.45.1", features = ["macros", "rt-multi-thread", "time", "io-util", "net", "sync"], optional = true }
 
 # embassy-usb-v0_4
-embassy-usb     = { version = "0.4", optional = true }
-embassy-futures = { version = "0.1.1", optional = true }
-embassy-time    = { version = "0.4.0", optional = true }
-defmt           = { version = "1.0.0", optional = true }
+defmt           = { version = "1.0.0",  optional = true }
+embassy-usb     = { version = "0.4",    optional = true }
+embassy-futures = { version = "0.1.1",  optional = true }
+embassy-time    = { version = "0.4.0",  optional = true }
+static_cell     = { version = "2.1",    optional = true}
 
 # nusb-v0_1
 nusb    = { version = "0.1.14", optional = true }
