+++ crates/ergot-base/src/interface_manager/eusb_0_4_client.rs
@@ -16,21 +16,30 @@ use crate::{
 use bbq2::{
     prod_cons::framed::FramedConsumer,
     queue::BBQueue,
-    traits::{coordination::cas::AtomicCoord, notifier::maitake::MaiNotSpsc, storage::Inline},
+    traits::{coordination::Coord, notifier::maitake::MaiNotSpsc, storage::Inline},
 };
+use core::sync::atomic::{AtomicU8, Ordering};
 use defmt::{debug, info, warn};
 use embassy_futures::select::{Either, select};
 use embassy_time::Timer;
-use embassy_usb::driver::{Driver, Endpoint, EndpointError, EndpointIn, EndpointOut};
+use embassy_usb::{
+    Builder, UsbDevice,
+    driver::{Driver, Endpoint, EndpointError, EndpointIn, EndpointOut},
+    msos::{self, windows_version},
+};
 use mutex::ScopedRawMutex;
+use static_cell::ConstStaticCell;
 
 /// An Embassy USB Bulk Transfer Interface Manager
 ///
 /// Only suitable for cases where this is an "edge" node that will not forward
 /// messages onwards. Assumes we only have a single upstream connection.
 #[derive(Default)]
-pub struct EmbassyUsbManager<const N: usize> {
-    inner: Option<EmbassyUsbManagerInner<N>>,
+pub struct EmbassyUsbManager<const N: usize, C>
+where
+    C: Coord + 'static,
+{
+    inner: Option<EmbassyUsbManagerInner<N, C>>,
     seq_no: u16,
 }
 
@@ -39,16 +48,24 @@ pub struct EmbassyUsbManager<const N: usize> {
 /// This manages the receiver operations, as well as manages the connection state.
 ///
 /// The `N` const generic buffer is the size of the outgoing buffer.
-pub struct Receiver<R: ScopedRawMutex + 'static, D: Driver<'static>, const N: usize> {
-    bbq: &'static BBQueue<Inline<N>, AtomicCoord, MaiNotSpsc>,
-    stack: &'static NetStack<R, EmbassyUsbManager<N>>,
+pub struct Receiver<R, D, const N: usize, C>
+where
+    R: ScopedRawMutex + 'static,
+    D: Driver<'static>,
+    C: Coord + 'static,
+{
+    bbq: &'static BBQueue<Inline<N>, C, MaiNotSpsc>,
+    stack: &'static NetStack<R, EmbassyUsbManager<N, C>>,
     rx: D::EndpointOut,
     net_id: Option<u16>,
 }
 
 /// Inner item visible when we DO have an active connection
-struct EmbassyUsbManagerInner<const N: usize> {
-    interface: ProducerHandle<N>,
+struct EmbassyUsbManagerInner<const N: usize, C>
+where
+    C: Coord + 'static,
+{
+    interface: ProducerHandle<N, C>,
     net_id: u16,
 }
 
@@ -64,13 +81,19 @@ enum ReceiverError {
     ConnectionClosed,
 }
 
-struct ProducerHandle<const N: usize> {
-    skt_tx: framed_stream::Interface<&'static BBQueue<Inline<N>, AtomicCoord, MaiNotSpsc>>,
+struct ProducerHandle<const N: usize, C>
+where
+    C: Coord + 'static,
+{
+    skt_tx: framed_stream::Interface<&'static BBQueue<Inline<N>, C, MaiNotSpsc>>,
 }
 
 // ---- impls ----
 
-impl<const N: usize> EmbassyUsbManager<N> {
+impl<const N: usize, C> EmbassyUsbManager<N, C>
+where
+    C: Coord + 'static,
+{
     pub const fn new() -> Self {
         Self {
             inner: None,
@@ -79,18 +102,24 @@ impl<const N: usize> EmbassyUsbManager<N> {
     }
 }
 
-impl<const N: usize> ConstInit for EmbassyUsbManager<N> {
+impl<const N: usize, C> ConstInit for EmbassyUsbManager<N, C>
+where
+    C: Coord + 'static,
+{
     #[allow(clippy::declare_interior_mutable_const)]
     const INIT: Self = Self::new();
 }
 
-impl<const N: usize> EmbassyUsbManager<N> {
+impl<const N: usize, C> EmbassyUsbManager<N, C>
+where
+    C: Coord + 'static,
+{
     fn common_send<'a, 'b>(
         &'b mut self,
         ihdr: &'a Header,
     ) -> Result<
         (
-            &'b mut EmbassyUsbManagerInner<N>,
+            &'b mut EmbassyUsbManagerInner<N, C>,
             CommonHeader,
             Option<&'a Key>,
         ),
@@ -167,7 +196,10 @@ impl<const N: usize> EmbassyUsbManager<N> {
     }
 }
 
-impl<const N: usize> InterfaceManager for EmbassyUsbManager<N> {
+impl<const N: usize, C> InterfaceManager for EmbassyUsbManager<N, C>
+where
+    C: Coord + 'static,
+{
     fn send<T: serde::Serialize>(
         &mut self,
         hdr: &Header,
@@ -213,11 +245,16 @@ impl<const N: usize> InterfaceManager for EmbassyUsbManager<N> {
     }
 }
 
-impl<R: ScopedRawMutex + 'static, D: Driver<'static>, const N: usize> Receiver<R, D, N> {
+impl<R, D, const N: usize, C> Receiver<R, D, N, C>
+where
+    R: ScopedRawMutex + 'static,
+    D: Driver<'static>,
+    C: Coord,
+{
     /// Create a new receiver object
     pub fn new(
-        q: &'static BBQueue<Inline<N>, AtomicCoord, MaiNotSpsc>,
-        stack: &'static NetStack<R, EmbassyUsbManager<N>>,
+        q: &'static BBQueue<Inline<N>, C, MaiNotSpsc>,
+        stack: &'static NetStack<R, EmbassyUsbManager<N, C>>,
         rx: D::EndpointOut,
     ) -> Self {
         Self {
@@ -423,13 +460,27 @@ impl<R: ScopedRawMutex + 'static, D: Driver<'static>, const N: usize> Receiver<R
     }
 }
 
+impl<R, D, const N: usize, C> Drop for Receiver<R, D, N, C>
+where
+    R: ScopedRawMutex + 'static,
+    D: Driver<'static>,
+    C: Coord,
+{
+    fn drop(&mut self) {
+        // No receiver? Drop the interface.
+        self.stack.with_interface_manager(|im| {
+            im.inner.take();
+        })
+    }
+}
+
 /// Transmitter worker task
 ///
 /// Takes a bbqueue from the NetStack of packets to send. While sending,
 /// we will timeout if
-pub async fn tx_worker<D: Driver<'static>, const N: usize>(
+pub async fn tx_worker<'a, D: Driver<'a>, const N: usize, C: Coord>(
     ep_in: &mut D::EndpointIn,
-    rx: FramedConsumer<&'static BBQueue<Inline<N>, AtomicCoord, MaiNotSpsc>>,
+    rx: FramedConsumer<&'static BBQueue<Inline<N>, C, MaiNotSpsc>>,
     timeout_ms_per_frame: usize,
     max_usb_frame_size: usize,
 ) {
@@ -475,7 +526,7 @@ pub async fn tx_worker<D: Driver<'static>, const N: usize>(
 }
 
 #[inline]
-async fn send_all<D>(
+async fn send_all<'a, D>(
     ep_in: &mut D::EndpointIn,
     out: &[u8],
     pending_frame: &mut bool,
@@ -483,7 +534,7 @@ async fn send_all<D>(
     max_usb_frame_size: usize,
 ) -> Result<(), TransmitError>
 where
-    D: Driver<'static>,
+    D: Driver<'a>,
 {
     if out.is_empty() {
         return Ok(());
@@ -523,5 +574,235 @@ where
     match select(send_fut, Timer::after_millis(timeout_ms as u64)).await {
         Either::First(res) => res,
         Either::Second(()) => Err(TransmitError::Timeout),
+    }
+}
+
+// Helper bits
+
+// ---- Constants ----
+
+pub const DEVICE_INTERFACE_GUIDS: &[&str] = &["{AFB9A6FB-30BA-44BC-9232-806CFC875321}"];
+/// Default time in milliseconds to wait for the completion of sending
+pub const DEFAULT_TIMEOUT_MS_PER_FRAME: usize = 2;
+/// Default max packet size for USB Full Speed
+pub const USB_FS_MAX_PACKET_SIZE: usize = 64;
+
+// ---- Statics ----
+
+/// Statically store our packet buffers
+static STINDX: AtomicU8 = AtomicU8::new(0xFF);
+static HDLR: ConstStaticCell<ErgotHandler> = ConstStaticCell::new(ErgotHandler {});
+
+// ---- Types ----
+
+pub struct UsbDeviceBuffers<
+    const CONFIG: usize = 256,
+    const BOS: usize = 256,
+    const CONTROL: usize = 64,
+    const MSOS: usize = 256,
+> {
+    /// Config descriptor storage
+    pub config_descriptor: [u8; CONFIG],
+    /// BOS descriptor storage
+    pub bos_descriptor: [u8; BOS],
+    /// CONTROL endpoint buffer storage
+    pub control_buf: [u8; CONTROL],
+    /// MSOS descriptor buffer storage
+    pub msos_descriptor: [u8; MSOS],
+}
+
+/// A helper type for `static` storage of buffers and driver components
+pub struct WireStorage<
+    const CONFIG: usize = 256,
+    const BOS: usize = 256,
+    const CONTROL: usize = 64,
+    const MSOS: usize = 256,
+> {
+    /// Usb buffer storage
+    pub bufs_usb: ConstStaticCell<UsbDeviceBuffers<CONFIG, BOS, CONTROL, MSOS>>,
+    // /// WireTx/Sender static storage
+    // pub cell: StaticCell<Mutex<M, EUsbWireTxInner<D>>>,
+}
+
+struct ErgotHandler {}
+
+// ---- impls ----
+
+// impl UsbDeviceBuffers
+
+impl<const CONFIG: usize, const BOS: usize, const CONTROL: usize, const MSOS: usize>
+    UsbDeviceBuffers<CONFIG, BOS, CONTROL, MSOS>
+{
+    /// Create a new, empty set of buffers
+    pub const fn new() -> Self {
+        Self {
+            config_descriptor: [0u8; CONFIG],
+            bos_descriptor: [0u8; BOS],
+            msos_descriptor: [0u8; MSOS],
+            control_buf: [0u8; CONTROL],
+        }
+    }
+}
+
+impl<const CONFIG: usize, const BOS: usize, const CONTROL: usize, const MSOS: usize> Default
+    for UsbDeviceBuffers<CONFIG, BOS, CONTROL, MSOS>
+{
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+// impl WireStorage
+
+impl<const CONFIG: usize, const BOS: usize, const CONTROL: usize, const MSOS: usize>
+    WireStorage<CONFIG, BOS, CONTROL, MSOS>
+{
+    /// Create a new, uninitialized static set of buffers
+    pub const fn new() -> Self {
+        Self {
+            bufs_usb: ConstStaticCell::new(UsbDeviceBuffers::new()),
+            // cell: StaticCell::new(),
+        }
+    }
+
+    /// Initialize the static storage, reporting as ergot compatible
+    ///
+    /// This must only be called once.
+    pub fn init_ergot<D: Driver<'static> + 'static>(
+        &'static self,
+        driver: D,
+        config: embassy_usb::Config<'static>,
+    ) -> (UsbDevice<'static, D>, D::EndpointIn, D::EndpointOut) {
+        let bufs = self.bufs_usb.take();
+
+        let mut builder = Builder::new(
+            driver,
+            config,
+            &mut bufs.config_descriptor,
+            &mut bufs.bos_descriptor,
+            &mut bufs.msos_descriptor,
+            &mut bufs.control_buf,
+        );
+
+        // Register a ergot-compatible string handler
+        let hdlr = HDLR.take();
+        builder.handler(hdlr);
+
+        // Add the Microsoft OS Descriptor (MSOS/MOD) descriptor.
+        // We tell Windows that this entire device is compatible with the "WINUSB" feature,
+        // which causes it to use the built-in WinUSB driver automatically, which in turn
+        // can be used by libusb/rusb software without needing a custom driver or INF file.
+        // In principle you might want to call msos_feature() just on a specific function,
+        // if your device also has other functions that still use standard class drivers.
+        builder.msos_descriptor(windows_version::WIN8_1, 0);
+        builder.msos_feature(msos::CompatibleIdFeatureDescriptor::new("WINUSB", ""));
+        builder.msos_feature(msos::RegistryPropertyFeatureDescriptor::new(
+            "DeviceInterfaceGUIDs",
+            msos::PropertyData::RegMultiSz(DEVICE_INTERFACE_GUIDS),
+        ));
+
+        // Add a vendor-specific function (class 0xFF), and corresponding interface,
+        // that uses our custom handler.
+        let mut function = builder.function(0xFF, 0, 0);
+        let mut interface = function.interface();
+        let stindx = interface.string();
+        STINDX.store(stindx.0, Ordering::Relaxed);
+        let mut alt = interface.alt_setting(0xFF, 0xCA, 0x7D, Some(stindx));
+        let ep_out = alt.endpoint_bulk_out(64);
+        let ep_in = alt.endpoint_bulk_in(64);
+        drop(function);
+
+        // Build the builder.
+        let usb = builder.build();
+
+        (usb, ep_in, ep_out)
+    }
+
+    /// Initialize the static storage.
+    ///
+    /// This must only be called once.
+    pub fn init<D: Driver<'static> + 'static>(
+        &'static self,
+        driver: D,
+        config: embassy_usb::Config<'static>,
+    ) -> (UsbDevice<'static, D>, D::EndpointIn, D::EndpointOut) {
+        let (builder, wtx, wrx) = self.init_without_build(driver, config);
+        let usb = builder.build();
+        (usb, wtx, wrx)
+    }
+    /// Initialize the static storage, without building `Builder`
+    ///
+    /// This must only be called once.
+    pub fn init_without_build<D: Driver<'static> + 'static>(
+        &'static self,
+        driver: D,
+        config: embassy_usb::Config<'static>,
+    ) -> (Builder<'static, D>, D::EndpointIn, D::EndpointOut) {
+        let bufs = self.bufs_usb.take();
+
+        let mut builder = Builder::new(
+            driver,
+            config,
+            &mut bufs.config_descriptor,
+            &mut bufs.bos_descriptor,
+            &mut bufs.msos_descriptor,
+            &mut bufs.control_buf,
+        );
+
+        // Add the Microsoft OS Descriptor (MSOS/MOD) descriptor.
+        // We tell Windows that this entire device is compatible with the "WINUSB" feature,
+        // which causes it to use the built-in WinUSB driver automatically, which in turn
+        // can be used by libusb/rusb software without needing a custom driver or INF file.
+        // In principle you might want to call msos_feature() just on a specific function,
+        // if your device also has other functions that still use standard class drivers.
+        builder.msos_descriptor(windows_version::WIN8_1, 0);
+        builder.msos_feature(msos::CompatibleIdFeatureDescriptor::new("WINUSB", ""));
+        builder.msos_feature(msos::RegistryPropertyFeatureDescriptor::new(
+            "DeviceInterfaceGUIDs",
+            msos::PropertyData::RegMultiSz(DEVICE_INTERFACE_GUIDS),
+        ));
+
+        // Add a vendor-specific function (class 0xFF), and corresponding interface,
+        // that uses our custom handler.
+        let mut function = builder.function(0xFF, 0, 0);
+        let mut interface = function.interface();
+        let mut alt = interface.alt_setting(0xFF, 0, 0, None);
+        let ep_out = alt.endpoint_bulk_out(64);
+        let ep_in = alt.endpoint_bulk_in(64);
+        drop(function);
+
+        (builder, ep_in, ep_out)
+    }
+}
+
+impl<const CONFIG: usize, const BOS: usize, const CONTROL: usize, const MSOS: usize> Default
+    for WireStorage<CONFIG, BOS, CONTROL, MSOS>
+{
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+// impl EUsbWireTx
+// ...
+
+// impl EUsbWireRx
+// ...
+
+// impl ErgotHandler
+
+impl embassy_usb::Handler for ErgotHandler {
+    fn get_string(&mut self, index: embassy_usb::types::StringIndex, lang_id: u16) -> Option<&str> {
+        use embassy_usb::descriptor::lang_id;
+
+        let stindx = STINDX.load(Ordering::Relaxed);
+        if stindx == 0xFF {
+            return None;
+        }
+        if lang_id == lang_id::ENGLISH_US && index.0 == stindx {
+            Some("ergot")
+        } else {
+            None
+        }
     }
 }
