+++ demos/nrf52840-eusb/Cargo.toml
@@ -10,12 +10,12 @@ cortex-m            = { version = "0.7.7", features = ["inline-asm", "critical-s
 embassy-executor    = { version = "0.7.0", features = ["task-arena-size-32768", "arch-cortex-m", "executor-thread", "executor-interrupt", "defmt"] }
 embassy-nrf         = { version = "0.3.1", features = ["defmt", "nrf52840", "time-driver-rtc1", "gpiote", "unstable-pac", "time"] }
 embassy-time        = { version = "0.4.0", features = ["defmt", "defmt-timestamp-uptime"] }
+mutex               = { version = "1.0.2", features = ["impl-unsafe-cortex-m-single-core"] }
 panic-probe         = { version = "1.0.0",   features = ["print-defmt"] }
 
 cortex-m-rt     = "0.7.5"
 defmt           = "1.0.1"
 defmt-rtt       = "1.0.0"
-mutex           = "1.0.0"
 postcard-rpc    = "0.11.9"
 static_cell = "2.1.1"
 embassy-usb = "0.4.0"
